% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buff_surf.R
\name{buff_surf}
\alias{buff_surf}
\title{Buffer a point along a terrain surface.}
\usage{
buff_surf(
  dtm,
  pnts,
  .dist = 30,
  .res = 1,
  .smooth = FALSE,
  .method = c("ksmooth", "chaikin", "spline"),
  ...
)
}
\arguments{
\item{dtm}{A SpatRaster Digital Terrain Model}

\item{pnts}{A SpatVector of points}

\item{.dist}{Numeric - the buffer distance.}

\item{.res}{Numeric - approximate resampling resolution of the (internally
calculated) cost raster.}

\item{.smooth}{Logical default FALSE. Should the output buffer be smoothed using \code{smoothr::smooth}?}

\item{.method}{The smoothing method to use - see \code{smoothr::smooth} for details.}

\item{...}{Passed to \code{smoothr::smooth} for eg. to control smoothness.}
}
\description{
Buffer a point along a terrain surface.
}
